pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genai-interview-qa"
version = "0.1.0"
description = "Generative AI Project for creating interview questions and answers from PDF documents"
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = {text = "MIT"}
authors = [
    {name = "Rohit Bharti", email = "rohit.bharti8211@gmail.com"}
]
maintainers = [
    {name = "Rohit Bharti", email = "rohit.bharti8211@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "langchain==0.3.25",
    "langchain-community==0.3.24",
    "langchain-core==0.3.61",
    "langchain-openai==0.3.18",
    "openai==1.82.0",
    "pypdf==5.5.0",
    "tiktoken==0.9.0",
    "aiofiles==24.1.0",
    "fastapi==0.115.9",
    "uvicorn==0.34.2",
    "jinja2==3.1.6",
    "python-multipart==0.0.20",
    "PyPDF2==3.0.1",
    "faiss-cpu==1.11.0",
    "python-dotenv==1.1.0",
]

[project.urls]
"Homepage" = "https://github.com/rbi-international/GenAI_Interview_Questions_using_LLMs"
"Bug Tracker" = "https://github.com/rbi-international/GenAI_Interview_Questions_using_LLMs/issues"
"Repository" = "https://github.com/rbi-international/GenAI_Interview_Questions_using_LLMs"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
static = ["**/*"]
templates = ["**/*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312", "py313"]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest]
testpaths = ["tests"]